fixed some warnings

+++ Source/ArtificialIntelligence/AIManager.cpp
@@ -105,7 +105,7 @@ void AIManager::update(float fTimeInSeconds)
         }
         float moveX = a.actionsToTake[AIComponent::eAction::ACTION_MOVE_RIGHT_LEFT];
         float moveY = a.actionsToTake[AIComponent::eAction::ACTION_MOVE_FORWARDS_BACKWARDS];
-        if ((moveX != 0) || (moveY != 0)) {
+        if (((int)moveX != 0) || ((int)moveY != 0)) {
             if (bot->isInControl()) {
                 physics->moveGlobal(moveX, moveY);
             }
